Split login_page.dart in login_page_content_mobile.dart and login_page_content_desktop.dart
