+++ lib/pages/mobile/login_page_content_mobile.dart
@@ -1,0 +1,24 @@
+import 'package:aspdm_project/widgets/login_form.dart';
+import 'package:flutter/material.dart';
+
+class LoginPageContentMobile extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Card(
+        margin: const EdgeInsets.all(25.0),
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Placeholder(fallbackHeight: 128.0),
+              SizedBox(height: 8.0),
+              LoginForm(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
